restore 'hooks-async.js' and set the '.test.js' extension to run it like others; remove its tests from 'hooks.test.js'
